fix: potential cycle because of floating point imprecition, use an external library to use fast robust geometric predicates (libigl-predicates)

+++ include/slot_array.hpp
@@ -9,28 +9,28 @@
 
 typedef struct {
     void* data;
-    size_t dataLen;
-    size_t dataElementSize;
+    uint32_t dataLen;
+    uint32_t dataElementSize;
 
     // Reading
     bool* active;
 
     // Writing
-    size_t* freeIndices;
-    size_t freeLen;
+    uint32_t* freeIndices;
+    uint32_t freeLen;
 
-    size_t capacity;
+    uint32_t capacity;
 } SlotArray;
 
-int SlotArrayInit(SlotArray* arr, size_t capacity, size_t elementSize);
+int SlotArrayInit(SlotArray* arr, uint32_t capacity, uint32_t elementSize);
 void SlotArrayTerminate(SlotArray* arr);
 void SlotArrayClear(SlotArray* arr);
-void* SlotArrayGet(const SlotArray* arr, size_t index);
+void* SlotArrayGet(const SlotArray* arr, uint32_t index);
 void* SlotArrayInsert(SlotArray* arr);
-void* SlotArrayInsertGetIndex(SlotArray* arr, size_t* index);
-void SlotArrayDel(SlotArray* arr, size_t index);
+void* SlotArrayInsertGetIndex(SlotArray* arr, uint32_t* index);
+void SlotArrayDel(SlotArray* arr, uint32_t index);
 bool SlotArrayFull(const SlotArray* arr);
 bool SlotArrayEmpty(const SlotArray* arr);
-size_t SlotArrayLen(const SlotArray* arr);
+uint32_t SlotArrayLen(const SlotArray* arr);
 
 #endif
