+++ src/slot_array.cpp
@@ -1,6 +1,6 @@
 #include "../include/slot_array.hpp"
 
-int SlotArrayInit(SlotArray* arr, size_t capacity, size_t dataElementSize) {
+int SlotArrayInit(SlotArray* arr, uint32_t capacity, uint32_t dataElementSize) {
     arr->data        = NULL;
     arr->active      = NULL;
     arr->freeIndices = NULL;
@@ -13,13 +13,13 @@ int SlotArrayInit(SlotArray* arr, size_t capacity, size_t dataElementSize) {
     arr->active          = (bool*) malloc(sizeof(bool) * capacity);
     if (!arr->active)
         goto error;
-    arr->freeIndices = (size_t*) malloc(sizeof(size_t) * capacity);
+    arr->freeIndices = (uint32_t*) malloc(sizeof(uint32_t) * capacity);
     if (!arr->freeIndices)
         goto error;
     arr->freeLen  = 0;
     arr->capacity = capacity;
 
-    for (size_t i = 0; i < arr->capacity; i++)
+    for (uint32_t i = 0; i < arr->capacity; i++)
         arr->active[i] = false;
     return EXIT_SUCCESS;
 error:
@@ -40,14 +40,14 @@ void SlotArrayTerminate(SlotArray* arr) {
 void SlotArrayClear(SlotArray* arr) {
     arr->dataLen = 0;
     arr->freeLen = 0;
-    for (size_t i = 0; i < arr->capacity; i++)
+    for (uint32_t i = 0; i < arr->capacity; i++)
         arr->active[i] = false;
 }
-void* SlotArrayGet(const SlotArray* arr, size_t index) {
+void* SlotArrayGet(const SlotArray* arr, uint32_t index) {
     return (uint8_t*) arr->data + index * arr->dataElementSize;
 }
 void* SlotArrayInsert(SlotArray* arr) {
-    ssize_t index;
+    uint32_t index;
 
     if (arr->freeLen == 0) {
         if (arr->dataLen >= arr->capacity)
@@ -58,7 +58,7 @@ void* SlotArrayInsert(SlotArray* arr) {
     arr->active[index] = true;
     return SlotArrayGet(arr, index);
 }
-void* SlotArrayInsertGetIndex(SlotArray* arr, size_t* index) {
+void* SlotArrayInsertGetIndex(SlotArray* arr, uint32_t* index) {
     if (arr->freeLen == 0) {
         if (arr->dataLen >= arr->capacity)
             return NULL;
@@ -68,7 +68,7 @@ void* SlotArrayInsertGetIndex(SlotArray* arr, size_t* index) {
     arr->active[*index] = true;
     return SlotArrayGet(arr, *index);
 }
-void SlotArrayDel(SlotArray* arr, size_t index) {
+void SlotArrayDel(SlotArray* arr, uint32_t index) {
     if (!arr->active[index])
         return;
     arr->active[index]               = false;
@@ -80,6 +80,6 @@ bool SlotArrayFull(const SlotArray* arr) {
 bool SlotArrayEmpty(const SlotArray* arr) {
     return arr->dataLen == arr->freeLen;
 }
-size_t SlotArrayLen(const SlotArray* arr) {
+uint32_t SlotArrayLen(const SlotArray* arr) {
     return arr->dataLen - arr->freeLen;
 }
